ADD feat beta sistema terminado
