+++ Dockerfile
@@ -9,7 +9,10 @@ RUN apt-get update && apt-get install -y \
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
+# Copy application code
 COPY app.py .
+COPY app/ app/
+COPY api/ api/
 COPY logging_config.py .
 COPY litellm_config.json .
 # Include tests and pytest config inside the image for in-container test runs
@@ -18,5 +21,5 @@ COPY tests/ tests/
 
 EXPOSE 8000 4000
 
-# Use python directly instead of shell script
+# Start via app.py to preserve existing scripts and avoid module name conflicts
 CMD ["python", "app.py"]
