+++ requirements.txt
@@ -3,6 +3,7 @@ uvicorn[standard]==0.32.1
 openai>=1.99.5
 httpx==0.28.1
 pydantic>=2.10.0
+pydantic-settings>=2.6.1
 python-multipart==0.0.20
 litellm
 langfuse
